+++ nucleus/__init__.py
@@ -171,7 +171,6 @@ from .model import Model
 from .model_run import ModelRun
 from .payload_constructor import (
     construct_annotation_payload,
-    construct_append_payload,
     construct_box_predictions_payload,
     construct_model_creation_payload,
     construct_segmentation_payload,
@@ -190,7 +189,6 @@ from .quaternion import Quaternion
 from .retry_strategy import RetryStrategy
 from .scene import Frame, LidarScene, VideoScene
 from .slice import Slice
-from .upload_response import UploadResponse
 from .utils import create_items_from_folder_crawl
 from .validate import Validate
 
@@ -604,19 +602,6 @@ class NucleusClient:
     def delete_dataset_item(self, dataset_id: str, reference_id) -> dict:
         dataset = self.get_dataset(dataset_id)
         return dataset.delete_item(reference_id)
-
-    @deprecated("Use Dataset.append instead.")
-    def populate_dataset(
-        self,
-        dataset_id: str,
-        dataset_items: List[DatasetItem],
-        batch_size: int = 20,
-        update: bool = False,
-    ):
-        dataset = self.get_dataset(dataset_id)
-        return dataset.append(
-            dataset_items, batch_size=batch_size, update=update
-        )
 
     @deprecated(msg="Use Dataset.ingest_tasks instead")
     def ingest_tasks(self, dataset_id: str, payload: dict):